Create isLegal
it is a method named isLegalNumber that takes two arguments. The first argument is an array 
whose elements are the digits of the number to test. The second argument is the base of the 
number represented by the first argument. The method returns 1 if the number represented by the 
array is a legal number in the given base, otherwise it returns 0.